+++ src/plonemeeting/restapi/serializer/item.py
@@ -35,7 +35,7 @@ class SerializeItemToJsonBase(object):
         """ """
         result["extra_include_linked_items"] = []
         # get the modes so we know which linked items to return
-        modes = self._get_param('mode', ["auto"], extra_include_name="linked_items")
+        modes = self.get_param('mode', ["auto"], extra_include_name="linked_items")
         linked_items = []
         for mode in modes:
             # every manually linked items
@@ -58,7 +58,7 @@ class SerializeItemToJsonBase(object):
             elif mode == "every_successors":
                 linked_items += self.context.get_every_successors(unrestricted=False)
         # now apply extra_include_linked_items_filters
-        filters = self._get_param("filter", [], extra_include_name="linked_items")
+        filters = self.get_param("filter", [], extra_include_name="linked_items")
         if filters:
             linked_items = filter_data(filters, linked_items)
         result["extra_include_linked_items"] = [